# config/swagger.yaml
openapi: 3.0.0
info:
  title: Task Manager API
  version: 1.0.0
  description: API documentation for the Task Manager project
servers:
  - url: http://localhost:3000
    description: Development server
paths:
  /tasks:
    get:
      summary: Get all tasks (paginated, sortable)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high]
          description: Sort by priority
        - in: query
          name: timestamp
          schema:
            type: string
          description: Sort by timestamp
        - in: query
          name: completed
          schema:
            type: boolean
          description: Sort by completed status
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
          description: Sort order
      responses:
        200:
          description: A paginated list of tasks
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - priority
              properties:
                title:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
                priority:
                  type: string
                  enum: [low, medium, high]
      responses:
        201:
          description: Task created
        400:
          description: Invalid input
  /tasks/{id}:
    get:
      summary: Get a task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Task found
        404:
          description: Task not found
    put:
      summary: Update a task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
                priority:
                  type: string
                  enum: [low, medium, high]
      responses:
        200:
          description: Task updated
        400:
          description: Invalid input
        404:
          description: Task not found
    delete:
      summary: Delete a task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Task deleted
        404:
          description: Task not found